requirements.txt
annotated-types==0.6.0
anyio==4.1.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
outcome==1.3.0.post0
packaging==23.2
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
PySocks==1.7.1
python-dotenv==1.0.0
python-telegram-bot==20.7
requests==2.31.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
